Change Includemod to work on lazy modtypes
Includemod now works directly on lazy modtypes & friends. The main
benefits are now longer having to force when getting module types from
Env.t, not having to copy things when storing them in Env.t and not
having to convert back and forth when strengthening.

This is a prerequisite for a subsequent change which will introduce lazy
value_descriptions.